+++ src/main/java/net/redflower/pridecraft/WorldGen/PrideCraftBiomeModifiers.java
@@ -15,6 +15,7 @@ import net.redflower.pridecraft.PrideCraft;
 public class PrideCraftBiomeModifiers {
 
     public static final ResourceKey<BiomeModifier> ADD_UNKNOWN_ORE = registerKey("add_unknown_ore");
+    public static final ResourceKey<BiomeModifier> ADD_PRIDE_ORE = registerKey("add_pride_ore");
 
     public static void bootstrap(BootstrapContext<BiomeModifier> context) {
         var placedFeatures = context.lookup(Registries.PLACED_FEATURE);
@@ -24,9 +25,14 @@ public class PrideCraftBiomeModifiers {
                 biomes.getOrThrow(BiomeTags.IS_OVERWORLD),
                 HolderSet.direct(placedFeatures.getOrThrow(PrideCraftPlacedFeatures.UNKNOWN_ORE_PLACED_KEY)),
                 GenerationStep.Decoration.UNDERGROUND_ORES));
+
+        context.register(ADD_PRIDE_ORE, new BiomeModifiers.AddFeaturesBiomeModifier(
+                biomes.getOrThrow(BiomeTags.IS_OVERWORLD),
+                HolderSet.direct(placedFeatures.getOrThrow(PrideCraftPlacedFeatures.PRIDE_ORE_PLACED_KEY)),
+                GenerationStep.Decoration.UNDERGROUND_ORES));
     }
 
-    //Created by Kaupenjoe
+
     private static ResourceKey<BiomeModifier> registerKey(String name) {
         return ResourceKey.create(NeoForgeRegistries.Keys.BIOME_MODIFIERS, ResourceLocation.fromNamespaceAndPath(PrideCraft.MODID, name));
     }