+++ src/main/java/net/redflower/pridecraft/item/custom/PrideFlowerItem.java
@@ -1,0 +1,11 @@
+package net.redflower.pridecraft.item.custom;
+
+import net.minecraft.world.item.ItemNameBlockItem;
+import net.minecraft.world.level.block.Block;
+
+public class PrideFlowerItem extends ItemNameBlockItem {
+
+    public PrideFlowerItem(Block block, Properties properties) {
+        super(block, properties);
+    }
+}